Spanner: Shifted postgres samples and tests to a separate dir
