+++ spanner/composer.json
@@ -1,5 +1,5 @@
 {
     "require": {
-        "google/cloud-spanner": "^1.49.0"
+        "google/cloud-spanner": "^1.48.0"
     }
 }
