+++ spanner/postgres/composer.json
@@ -1,0 +1,5 @@
+{
+    "require": {
+        "google/cloud-spanner": "^1.49.0"
+    }
+}
