+++ spanner/postgres/src/pg_create_database.php
@@ -21,7 +21,7 @@
  * @see https://github.com/GoogleCloudPlatform/php-docs-samples/tree/master/spanner/README.md
  */
 
-namespace Google\Cloud\Samples\Spanner;
+namespace Google\Cloud\Samples\Spanner\Postgres;
 
 // [START spanner_create_postgres_database]
 use Google\Cloud\Spanner\SpannerClient;
@@ -78,5 +78,5 @@ function pg_create_database(string $instanceId, string $databaseId): void
 // [END spanner_create_postgres_database]
 
 // The following 2 lines are only needed to run the samples
-require_once __DIR__ . '/../../testing/sample_helpers.php';
+require_once __DIR__ . '/../../../testing/sample_helpers.php';
 \Google\Cloud\Samples\execute_sample(__FILE__, __NAMESPACE__, $argv);
