+++ spanner/postgres/src/pg_dml_getting_started_update.php
@@ -21,7 +21,7 @@
  * @see https://github.com/GoogleCloudPlatform/php-docs-samples/tree/master/spanner/README.md
  */
 
-namespace Google\Cloud\Samples\Spanner;
+namespace Google\Cloud\Samples\Spanner\Postgres;
 
 // [START spanner_postgresql_dml_getting_started_update]
 use Google\Cloud\Spanner\SpannerClient;
@@ -95,5 +95,5 @@ function pg_dml_getting_started_update(string $instanceId, string $databaseId): 
 // [END spanner_postgresql_dml_getting_started_update]
 
 // The following 2 lines are only needed to run the samples
-require_once __DIR__ . '/../../testing/sample_helpers.php';
+require_once __DIR__ . '/../../../testing/sample_helpers.php';
 \Google\Cloud\Samples\execute_sample(__FILE__, __NAMESPACE__, $argv);
