+++ spanner/postgres/src/pg_interleaved_table.php
@@ -21,7 +21,7 @@
  * @see https://github.com/GoogleCloudPlatform/php-docs-samples/tree/master/spanner/README.md
  */
 
-namespace Google\Cloud\Samples\Spanner;
+namespace Google\Cloud\Samples\Spanner\Postgres;
 
 // [START spanner_postgresql_interleaved_table]
 use Google\Cloud\Spanner\SpannerClient;
@@ -68,5 +68,5 @@ function pg_interleaved_table(string $instanceId, string $databaseId, string $pa
 // [END spanner_postgresql_interleaved_table]
 
 // The following 2 lines are only needed to run the samples
-require_once __DIR__ . '/../../testing/sample_helpers.php';
+require_once __DIR__ . '/../../../testing/sample_helpers.php';
 \Google\Cloud\Samples\execute_sample(__FILE__, __NAMESPACE__, $argv);
