--- spanner/src/pg_functions.php
@@ -1,55 +1,0 @@
-<?php
-/**
- * Copyright 2022 Google Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * For instructions on how to run the full sample:
- *
- * @see https://github.com/GoogleCloudPlatform/php-docs-samples/tree/master/spanner/README.md
- */
-
-namespace Google\Cloud\Samples\Spanner;
-
-// [START spanner_postgresql_functions]
-use Google\Cloud\Spanner\SpannerClient;
-
-/**
- * Call a server side function on a Spanner PostgreSQL database
- *
- * @param string $instanceId The Spanner instance ID.
- * @param string $databaseId The Spanner database ID.
- */
-function pg_functions(string $instanceId, string $databaseId): void
-{
-    $spanner = new SpannerClient();
-    $instance = $spanner->instance($instanceId);
-    $database = $instance->database($databaseId);
-
-    // Use the PostgreSQL `to_timestamp` function to convert a number of
-    // seconds since epoch to a timestamp.
-    // 1284352323 seconds = Monday, September 13, 2010 4:32:03 AM.
-    $results = $database->execute('SELECT to_timestamp(1284352323) AS time');
-
-    $row = $results->rows()->current();
-    $time = $row['time'];
-
-    printf('1284352323 seconds after epoch is %s' . PHP_EOL, $time);
-}
-// [END spanner_postgresql_functions]
-
-// The following 2 lines are only needed to run the samples
-require_once __DIR__ . '/../../testing/sample_helpers.php';
-\Google\Cloud\Samples\execute_sample(__FILE__, __NAMESPACE__, $argv);
